Add process managment and move all files from server new place
