+++ docker-compose.prod.yml
@@ -1,0 +1,11 @@
+version: '3.1'
+
+services:
+  node:
+    build:
+      context: .
+      args:
+        - NODE_ENV=production
+    command: yarn prod:docker
+    environment:
+      - NODE_ENV=production
