+++ docker-compose.yml
@@ -6,7 +6,7 @@ services:
       context: .
       args:
         - NODE_ENV=development
-    command: ./node_modules/nodemon/bin/nodemon.js --inspect=0.0.0.0:9229 server.js --exec babel-node
+    command: ./node_modules/nodemon/bin/nodemon.js --inspect=0.0.0.0:9229 src/server/server.js --exec babel-node
     ports:
       - "80:80"
       - "9229:9229"
