+++ src/server/setup/serverProductionSetup.js
@@ -1,7 +1,7 @@
 import express from 'express'
 import compression from 'compression'
-import { STATIC_PATH } from '../src/shared/config'
-import { isProd } from '../src/shared/util'
+import { STATIC_PATH } from '../../shared/config'
+import { isProd } from '../../shared/util'
 
 export default function (app) {
   if (isProd) {
